lab3/requirements.txt
pillow~=12.0.0
numpy~=2.3.4